requirements.txt
ollama==0.6.1
langchain==1.2.12
langchain-community==0.4.1
faiss-cpu==1.13.2
streamlit==1.55.0
pypdf==6.8.0
sentence-transformers==5.3.0
notebook==7.5.5
mcp[cli]==1.26.0
a2a-sdk==0.3.25
uvicorn==0.42.0
httpx==0.28.1
nest_asyncio2==1.7.2

